python/requirements.txt
hvac
google-auth
google-cloud-storage